VFS: Introduce inode-getting helpers for layered/unioned fs environments
From ed9ead2c17c1c2a4a46c674fa19a87347f354638 Mon Sep 17 00:00:00 2001
From: David Howells <dhowells@redhat.com>
Date: Thu, 29 Jan 2015 12:02:27 +0000
Subject: [PATCH] VFS: Introduce inode-getting helpers for layered/unioned fs environments

Introduce some function for getting the inode
(and also the dentry) in an environment where
layered/unioned filesystems are in operation.

The problem is that we have places where we need *both*
the union dentry and the lower source or workspace inode
or dentry available, but we can only have a handle on one
of them.  Therefore we need to derive the handle to the
other from that.

The idea is to introduce an extra field in struct dentry
that allows the union dentry to refer to and pin
the lower dentry.

Change-Id: I4cce731a75b4c3f47262f948b7d63180fb556acd
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Paul Keith <javelinanddart@gmail.com>
---